Work towards enabling index generation
Abstracted the index package format and the package's metadata.tar
format into `PackageCommon` to faciliate storing the required fields
into the DB to then, on the fly, generate the Package index for serving
to clients.

Also reworked the `.mise.toml` to add a new `mise run watch` command to
faciliate hot reloading of the server component.